pyproject.toml
[project]
name = "nse-universe"
version = "0.1.0"
description = "Point-in-time NSE top-500 custom index membership oracle for backtesting"
authors = [
    { name = "Rakesh Waghela", email = "javaindia@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.2",
    "curl-cffi>=0.15.0",
    "duckdb>=1.5.2",
    "fastapi>=0.136.0",
    "pandas>=3.0.2",
    "pyarrow>=23.0.1",
    "pyyaml>=6.0.3",
    "questionary>=2.1.1",
    "rich>=15.0.0",
    "tqdm>=4.67.3",
    "uvicorn[standard]>=0.44.0",
    "yfinance>=1.3.0",
]

[project.scripts]
nse-universe = "nse_universe.cli:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[dependency-groups]
dev = [
    "pillow>=12.2.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
